shouts/requirements.txt
django==3.1.4
django-extensions==3.0.9
djangorestframework==3.12.2